Remove unnecessary type conversions
Remove redundant uint64() and int32() conversions flagged by the
unconvert linter. GetNrPages() already returns uint64 and Config.Pid
is already int32.

+++ crit/mempages.go
@@ -235,14 +235,14 @@ func (mr *MemoryReader) SearchPattern(pattern string, escapeRegExpCharacters boo
 
 	for _, entry := range mr.pagemapEntries {
 		startAddr := entry.GetVaddr()
-		endAddr := startAddr + uint64(entry.GetNrPages())*uint64(mr.pageSize)
+		endAddr := startAddr + entry.GetNrPages()*uint64(mr.pageSize)
 
 		initialOffset := uint64(0)
 		for _, e := range mr.pagemapEntries {
 			if e == entry {
 				break
 			}
-			initialOffset += uint64(e.GetNrPages()) * uint64(mr.pageSize)
+			initialOffset += e.GetNrPages() * uint64(mr.pageSize)
 		}
 
 		for offset := uint64(0); offset < endAddr-startAddr; offset += uint64(chunkSize) {
